+++ dqt/iterable_settings.py
