+++ dqt/settings_menu.py
@@ -2,7 +2,7 @@ import sys
 from itertools import cycle
 from typing import List
 
-from dqt.iterableSettings import SubDictEnum, IterableSettings, Special
+from dqt.iterable_settings import SubDictEnum, IterableSettings, Special
 from dqt.ui_utils import invalid_choice, menu
 from dqt.styletext import StyleText as Txt, clear_console
