+++ dqt/tracker.py
@@ -8,7 +8,7 @@ from dqt.manager import Manager
 from dqt.stats import Stats
 from dqt.ui_utils import cont_on_enter, err, invalid_choice, menu
 from dqt.styletext import StyleText as Txt
-from dqt.settingsMenu import SettingsMenu
+from dqt.settings_menu import SettingsMenu
 
 from settings import CONFIGS
